+++ CHANGELOG.md
@@ -11,6 +11,14 @@
 ** gpfs_fs_inodes_free to gpfs_fs_free_inodes
 ** gpfs_fs_inodes_total to gpfs_fs_total_inodes
 ** gpfs_fs_inodes_used to gpfs_fs_used_inodes
+** gpfs_fs_total_inodes to gpfs_fs_inodes
+** gpfs_fs_total_bytes to gpfs_fs_size_bytes
+** gpfs_fs_metadata_total_bytes to gpfs_fs_metadata_size_bytes
+* Removed metrics that can be calculated using other metrics
+** gpfs_fs_metadata_free_percent
+** gpfs_fs_free_percent metrics
+* Remove nodename label from gpfs_perf_* metrics, replace with gpfs_perf_info metric
+* mmces and mmhealth status metrics will always have value 1, only the `status` label will change
 
 ### Improvements